+++ tinytga/src/pixels.rs
@@ -1,15 +1,25 @@
-use crate::RawPixels;
+use core::marker::PhantomData;
 use embedded_graphics::prelude::*;
 
-/// Iterator over individual TGA pixels
+use crate::RawPixels;
+
+/// Iterator over individual TGA pixels.
+///
+/// See the [`pixels`] method for additional information.
+///
+/// [`pixels`]: struct.Tga.html#method.pixels
 #[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug)]
 pub struct Pixels<'a, 'b, C> {
-    raw: RawPixels<'a, 'b, C>,
+    raw: RawPixels<'a, 'b>,
+    color_type: PhantomData<C>,
 }
 
 impl<'a, 'b, C> Pixels<'a, 'b, C> {
-    pub(crate) fn new(raw: RawPixels<'a, 'b, C>) -> Self {
-        Self { raw }
+    pub(crate) fn new(raw: RawPixels<'a, 'b>) -> Self {
+        Self {
+            raw,
+            color_type: PhantomData,
+        }
     }
 }