+++ tinytga/tests/errors.rs
@@ -1,15 +1,16 @@
 use embedded_graphics::{
     pixelcolor::{Gray8, Rgb888},
     prelude::*,
+    primitives::Rectangle,
 };
 use std::iter::repeat;
-use tinytga::{ParseError, RawPixel, Tga};
+use tinytga::{ParseError, RawPixel, RawTga, Tga};
 
 #[test]
 fn color_map() {
     // The color map in "error_color_map.tga" has too many entries and is larger than the file
     assert_eq!(
-        Tga::from_slice_raw(include_bytes!("../tests/error_color_map.tga")),
+        RawTga::from_slice(include_bytes!("../tests/error_color_map.tga")),
         Err(ParseError::ColorMap)
     );
 }
@@ -17,17 +18,16 @@ fn color_map() {
 #[test]
 fn image_data_missing() {
     // The image data in "error_no_image_data.tga" is missing
-    let tga = Tga::from_slice_raw(include_bytes!("../tests/error_no_image_data.tga")).unwrap();
+    let tga = RawTga::from_slice(include_bytes!("../tests/error_no_image_data.tga")).unwrap();
 
-    assert!(tga.raw_image_data().is_empty());
+    assert!(tga.image_data().is_empty());
 
-    let expected: Vec<_> = tga
-        .bounding_box()
+    let expected: Vec<_> = Rectangle::new(Point::zero(), tga.size())
         .points()
         .map(|p| RawPixel::new(p, 0))
         .collect();
 
-    let pixels: Vec<_> = tga.raw_pixels().collect();
+    let pixels: Vec<_> = tga.pixels().collect();
 
     assert_eq!(pixels, expected);
 }
@@ -36,18 +36,17 @@ fn image_data_missing() {
 fn image_data_truncated() {
     // The image data in "error_truncated_image_data.tga" is truncated.
     let tga =
-        Tga::from_slice_raw(include_bytes!("../tests/error_truncated_image_data.tga")).unwrap();
+        RawTga::from_slice(include_bytes!("../tests/error_truncated_image_data.tga")).unwrap();
 
-    assert_eq!(tga.raw_image_data(), &[1, 2, 3, 4, 5, 6, 7, 8]);
+    assert_eq!(tga.image_data(), &[1, 2, 3, 4, 5, 6, 7, 8]);
 
-    let expected: Vec<_> = tga
-        .bounding_box()
+    let expected: Vec<_> = Rectangle::new(Point::zero(), tga.size())
         .points()
         .zip((1..=8).chain(repeat(0)))
         .map(|(p, c)| RawPixel::new(p, c))
         .collect();
 
-    let pixels: Vec<_> = tga.raw_pixels().collect();
+    let pixels: Vec<_> = tga.pixels().collect();
 
     assert_eq!(pixels, expected);
 }